remove bind for apps dashboard
